added vllm file logging to call_ll
